requirements.txt
paramiko>=3.4.0
matplotlib>=3.5.0
requests>=2.25.0
urllib3>=1.26.0
psutil>=5.8.0
python-dotenv>=1.0.0
fpdf2>=2.7.0
PyMuPDF>=1.23.0
Pillow>=9.0.0
ttkbootstrap>=1.10.0
customtkinter>=5.2.0
